Extract shared login helper into e2e/utils/auth.cjs
The login flow was duplicated across login, storyEditor, and
createWorldModal specs. Centralise the UI login + credentials +
LOGIN_TIMEOUT constant into a shared utility so every spec stays in
sync (including the 20 s cold-start timeout).
